% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_search.R
\name{BootstrapTree}
\alias{BootstrapTree}
\title{Bootstrap tree search with inapplicable data}
\usage{
BootstrapTree(tree, morphyObj, maxIter, maxHits, verbosity = 1, ...)
}
\arguments{
\item{tree}{A tree of class \code{\link{phylo}}, with tip labels in the order generated by 
\code{\link{RenumberTips}}, i.e. corresponding corresponding the sequence of taxa 
in the corresponding Morphy object.}

\item{morphyObj}{A morphy object created with \code{\link{LoadMorphy}}}

\item{maxIter}{maximum number of iterations to perform in tree search}

\item{maxHits}{maximum number of hits to accomplish in tree search}

\item{verbosity}{Level of detail to display in console: larger numbers provides more verbose feedback to the user;}

\item{\dots}{further parameters to send to DoTreeSearch}
}
\value{
A tree that is optimal under a random sampling of the original characters
}
\description{
Bootstrap tree search with inapplicable data
}
