% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_search.R
\name{DoTreeSearch}
\alias{DoTreeSearch}
\title{DoTreeSearch}
\usage{
DoTreeSearch(tree, morphyObj, method = "NNI", maxiter = 100, maxhits = 20,
  forest.size = 1, cluster = NULL, verbosity = 1, criterion = NULL, ...)
}
\arguments{
\item{tree}{A tree of class \code{\link{phylo}}, with tip labels in the order generated by 
\code{\link{RenumberTips}}, i.e. corresponding corresponding the sequence of taxa 
in the corresponding Morphy object.}

\item{morphyObj}{A morphy object created with \code{\link{LoadMorphy}}}
}
\description{
Performs a tree search
}
\details{
Does the hard work of searching for a most parsimonious tree.
End-users are expected to access this function through its wrapper, TreeSearch
It is also called directly by Ratchet and Sectorial functions
}
\author{
Martin R. Smith
}
\keyword{internal}
