% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitch.R
\name{InapplicableFitch}
\alias{InapplicableFitch}
\title{Calculate parsimony score with inapplicable data}
\usage{
InapplicableFitch(tree, morphyData, detail = 1)
}
\arguments{
\item{tree}{A tree of class \code{phylo}}

\item{morphyData}{A \code{phyDat} or \code{morphyDat} object, perhaps generated with 
\code{\link{phyDat}} or \code{\link{MorphyData}}}

\item{detail}{Leave as 1 to just return parsimony score, or specify c(1, 2, 3) for additional detail (see below)}
}
\description{
Uses code modified from the Morphy library to calculate a parsimony score 
in datasets that contain inapplicable data
}
\examples{
data(SigSut)
taxa <- names(SigSut)
tree <- rtree(length(taxa), tip.label=taxa, br=NULL)
morphy <- MorphyDat(tree, SigSut.phy)

}
\seealso{
\code{\link{MorphyData}}
}
\author{
Martin Smith (using C code adapted from MorphyLib, author Martin Brazeau)
}
