% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_search.R
\name{InapplicableSectorial}
\alias{InapplicableSectorial}
\title{Sectorial Search with inapplicable data}
\usage{
InapplicableSectorial(tree, dataset, maxit = 100, maxiter = 500, k = 5,
  verbosity = 0, smallest.sector = 4, largest.sector = 1e+06,
  rearrangements = "NNI", criterion = NULL, ...)
}
\arguments{
\item{PARAM}{is a parameter you should send to it}
}
\value{
This function returns :
}
\description{
\code{InapplicableSectorial} does something useful
}
\examples{
to_do <- TRUE

}
\author{
Martin R. Smith
}
