% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{MorphyDat}
\alias{MorphyDat}
\title{PhyDat to MorphyDat}
\usage{
MorphyDat(phydat)
}
\arguments{
\item{phydat}{A \code{phyDat} object}
}
\description{
Converts a PhyDat object to allow processing by MorphyDat
}
\examples{
morphy <- MorphyDat(SigSut.phy)

}
\seealso{
\code{\link{phyDat}}
}
\author{
Martin Smith
}
