% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitch.R
\name{MorphyLength}
\alias{MorphyLength}
\title{Calculate parsimony score with inapplicable data}
\usage{
MorphyLength(tree, morphyObj)
}
\arguments{
\item{tree}{A tree of class \code{\link[ape]{phylo}}, with tip labels in the order generated by 
\code{\link[TreeSearch]{RenumberTips}}, i.e. corresponding to the sequence of taxa 
in the corresponding Morphy object.}

\item{morphyObj}{A morphy object created with \code{\link{LoadMorphy}}.}
}
\value{
The length of the tree (after weighting)
}
\description{
Calculate parsimony score with inapplicable data
}
\seealso{
LoadMorphy
}
\author{
Martin R. Smith
}
\keyword{internal}
