% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphyex.R
\name{mpl_set_parsim_t}
\alias{mpl_set_parsim_t}
\title{Sets a character's parsimony function type}
\usage{
mpl_set_parsim_t(char_id, tname = "typename", morphyobj)
}
\arguments{
\item{morphyobj}{An instance of the Morphy object.}

\item{charID}{The number of the character (transformation series) as defined
in the input matrix.  The first character is numbered 1 (one).}

\item{chtype}{The parsimony function type as defined in morphydefs.h}
}
\value{
A Morphy error code.
}
\description{
Set the parsimony function type to one defined in the
morphydefs.h header file. Setting the character to type NONE_T will also
cause it to be excluded from any further calculations.
}
\author{
Martin Brazeau
}
