% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_rearrange.R
\name{NNI}
\alias{NNI}
\title{Perform one NNI rearrangement at a given branch}
\usage{
NNI(tree)
}
\arguments{
\item{tree}{A tree of class \code{\link{phylo}}}
}
\value{
One of the two trees resulting when a NNI rearrangement is 
        performed at a random internal edge
}
\description{
Perform one NNI rearrangement at a given branch
}
