% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_search.R
\name{Ratchet}
\alias{Ratchet}
\alias{RatchetConsensus}
\title{Parsimony Ratchet}
\usage{
Ratchet(tree, dataset, keepAll = FALSE, outgroup = NULL, maxIt = 100,
  maxIter = 5000, maxHits = 40, k = 10, verbosity = 1,
  rearrangements = list(TreeSearch::RootedTBR, TreeSearch::RootedSPR,
  TreeSearch::RootedNNI), ...)

RatchetConsensus(tree, dataset, maxIt = 5000, maxIter = 500, maxHits = 20,
  k = 10, verbosity = 0, rearrangements = list(TreeSearch::NNI),
  nSearch = 10, ...)
}
\arguments{
\item{tree}{A tree of class \code{\link{phylo}}}

\item{dataset}{A phylogenetic data matrix of class \code{\link[phangorn]{phyDat}}, whose names 
correspond to the labels of any accompanying tree}

\item{keepAll}{Set to \code{TRUE} to report all MPTs encountered during the search, perhaps to analyze consensus}

\item{outgroup}{a vector specifying all tips in the outgroup; if unspecified then identical
trees with different roots will be considered unique;}

\item{maxIt}{maximum ratchet iterations to perform;}

\item{maxIter}{maximum rearrangements to perform on each bootstrap or ratchet iteration;}

\item{maxHits}{maximum times to hit best score before terminating a tree search within a pratchet iteration;}

\item{k}{stop when k ratchet iterations have found the same best score;}

\item{verbosity}{Level of detail to display in console: larger numbers provides more verbose feedback to the user;}

\item{rearrangements}{(list of) function(s) to use when rearranging trees
e.g. \code{list(TreeSearch::RootedTBR, TreeSearch::RootedNNI)}}

\item{\dots}{other arguments to pass to subsequent functions.}
}
\value{
This function returns a tree modified by parsimony ratchet iteration, retaining the position of the root.
}
\description{
\code{Ratchet} uses the parsimony ratchet (Nixon 1999) to search for a more parsimonious tree.
}
\section{Functions}{
\itemize{
\item \code{RatchetConsensus}: returns a list of optimal trees produced by nSearch Ratchet searches
}}

\examples{
{
data('Lobo')
Ratchet(RandomTree(Lobo.phy), dataset=Lobo.phy, outgroup='Tubiluchus_Priapulida', maxIt=1, maxIter=50)
}
}
\references{
Nixon, K. C. (1999). \cite{The Parsimony Ratchet, a new method for rapid parsimony analysis.}
 Cladistics, 15(4), 407-414. doi:\href{http://dx.doi.org/10.1111/j.1096-0031.1999.tb00277.x}{10.1111/j.1096-0031.1999.tb00277.x}
}
\seealso{
\code{\link{pratchet}}

\code{\link{TreeSearch}}

\code{\link{SectorialSearch}}
}
\author{
Martin R. Smith

Adapted from \code{\link[phangorn]{pratchet}} in the \pkg{phangorn} package, which does not preserve the position of the root.
}
\keyword{tree}
