% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_search.R
\name{RatchetConsensus}
\alias{RatchetConsensus}
\title{Consensus of Ratchet iterations}
\usage{
RatchetConsensus(tree, dataset, maxit = 5000, maxiter = 500, maxhits = 20,
  k = 10, verbosity = 0, rearrangements = "NNI", criterion = NULL,
  nSearch = 10, ...)
}
\arguments{
\item{PARAM}{is a parameter you should send to it}
}
\value{
This function returns :
}
\description{
\code{FUNCTIONNAME} does something useful but undocumented
}
\examples{
to_do <- TRUE

}
\author{
Martin R. Smith
}
