% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_search.R
\name{SectorialSearch}
\alias{SectorialSearch}
\title{Sectorial Search}
\usage{
SectorialSearch(tree, dataset, concavity = NULL, rearrangements = "NNI",
  maxIter = 2000, cluster = NULL, verbosity = 3, ...)
}
\arguments{
\item{tree}{a rooted, resolved tree in \code{\link{phylo}} format from which to start the search;}

\item{dataset}{A phylogenetic data matrix of class \code{\link[phangorn]{phyDat}}, whose names 
correspond to the labels of any accompanying tree}

\item{concavity}{concavity constant for implied weighting (not currently implemented!)}

\item{rearrangements}{method to use when rearranging subtrees: NNI, SPR or TBR;}

\item{maxIter}{maximum number of rearrangements to perform on each sectorial iteration;}

\item{cluster}{a cluster prepared using \code{\link{PrepareCluster}}; may speed up search on multicore machines;}

\item{verbosity}{Level of detail to display in console: larger numbers provides more verbose feedback to the user;}

\item{\dots}{other arguments to pass to subsequent functions.}

\item{outgroup}{a vector listing the taxa that form the outgroup;}

\item{maxIt}{maximum number of sectorial iterations to perform;}

\item{k}{stop when \code{k} searches have improved their sectorial score;}

\item{smallest.sector}{sectors with fewer than \code{smallest.sector} taxa will not be selected; \kbd{4} is the smallest sensible value;}

\item{largest.sector}{sectors with more than \code{largest.sector} taxa will not be selected;}
}
\value{
a rooted tree of class \code{phylo}.
}
\description{
\code{SectorialSearch} performs a sectorial search on a tree, preserving the position of the root.
}
\details{
\code{InapplicableSectorial} performs a sectorial search on the tree specified. A sectorial search 
detaches a random part of the tree, performs rearrangments on this subtree, then reattaches it 
to the main tree (Goloboff, 1999).
The improvement to local \var{pscore} hopefully (but not necessarily) improves the overall \var{pscore}.
As such, the output of \code{InapplicableSectorial} should be treated by further \acronym{TBR (/SPR/NNI)}
rearrangements and only retained if the ultimate parsimony score is better than 
that of the original tree.

\code{SectorialSearch} is a basic recipe that runs \code{InapplicableSectorial} followed by a few rounds
of tree rearrangement, returning a tree whose \var{pscore} is no worse than that of \code{start.tree}.
}
\examples{
require('ape')
data('SigSut')
outgroup <- c('Lingula', 'Mickwitzia', 'Neocrania')
njtree <- ape::root(nj(dist.hamming(SigSut.phy)), outgroup, resolve.root=TRUE)
njtree$edge.length <- NULL; njtree<-SetOutgroup(njtree, outgroup)
InapplicableSectorial(njtree, SigSut.phy, outgroup, maxIt=1, maxIter=50, largest.sector=7)
\dontrun{SectorialSearch(njtree, SigSut.phy, outgroup, 'SPR') # Will be time-consuming }


}
\references{
Goloboff, P. (1999). \cite{Analyzing large data sets in reasonable times: solutions for composite optima.} Cladistics, 15(4), 415-428. doi:\href{http://dx.doi.org/10.1006/clad.1999.0122}{10.1006/clad.1999.0122}
}
\seealso{
\code{\link{TreeSearch}}

\code{\link{Ratchet}}
}
\author{
Martin R. Smith
}
\keyword{tree}
