% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_rearrange.R
\name{RootedSPR}
\alias{RootedSPR}
\alias{SPR}
\title{Rooted SPR rearrangement}
\usage{
RootedSPR(tree)

SPR(tree)
}
\arguments{
\item{tree}{A tree of class \code{\link{phylo}}}
}
\description{
Rooted SPR rearrangement

Perform one \acronym{SPR} rearrangement on a tree
}
\section{Functions}{
\itemize{
\item \code{RootedSPR}: Perform \acronym{SPR} operation, retaining position of root
}}

