% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphy_objects.R
\name{summary.morphyPtr}
\alias{summary.morphyPtr}
\title{Details the attributes of a morphy object}
\usage{
\method{summary}{morphyPtr}(object, ...)
}
\arguments{
\item{object}{A Morphy object}
}
\value{
A list detailing the number of taxa, internal nodes, and characters and their weigths.
}
\description{
Details the attributes of a morphy object
}
\author{
Martin R. Smith
}
