% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_search.R
\name{TreeSearch}
\alias{TreeSearch}
\title{Search for most parsimonious trees}
\usage{
TreeSearch(tree, dataset, Rearrange = TreeSearch::RootedTBR, maxIter = 100,
  maxHits = 20, forestSize = 1, cluster = NULL, verbosity = 1, ...)
}
\arguments{
\item{tree}{a fully-resolved starting tree in \code{\link{phylo}} format, with the desired outgroup; 
edge lengths are not supported and will be deleted;}

\item{dataset}{A phylogenetic data matrix of class \code{\link[phangorn]{phyDat}}, whose names 
correspond to the labels of any accompanying tree}

\item{Rearrange}{Function used to rearrange trees; default: \code{\link{RootedTBR}}}

\item{maxIter}{the maximum number of iterations to perform before abandoning the search;}

\item{maxHits}{the maximum times to hit the best pscore before abandoning the search;}

\item{forestSize}{the maximum number of trees to return - useful in concert with \code{\link{consensus}};}

\item{cluster}{a cluster, prepared with code{link{PrepareCluster}}, to accelerate 
searches on multicore machines. #TODO not tested. #TODO restore code link}

\item{verbosity}{Level of detail to display in console: larger numbers provides more verbose feedback to the user;}

\item{\dots}{other arguments to pass to subsequent functions.}
}
\value{
{
This function returns a tree, with an attribute \code{pscore} conveying its parsimony score.
Note that the parsimony score will be inherited from the tree's attributes, which is only valid if it 
was generated using the same \code{data} that is passed here.
}
}
\description{
Run standard search algorithms (\acronym{NNI}, \acronym{SPR} or \acronym{TBR}) 
to search for a more parsimonious tree.
}
\examples{
library('ape'); library('phangorn')
data('inapplicable.datasets')
my.phyDat <- inapplicable.phyData[[1]]
outgroup <- names(my.phyDat)[1]
njtree <- ape::root(ape::nj(phangorn::dist.hamming(SigSut.phy)), outgroup, resolve.root=TRUE)
njtree$edge.length <- NULL
njtree <- ape::root(njtree, outgroup, resolve.root=TRUE)

\dontrun{
TreeSearch(njtree, my.phyDat, maxIter=20, Rearrange=TreeSearch::NNI)
TreeSearch(njtree, my.phyDat, maxIter=20, Rearrange=TreeSearch::RootedSPR)
}

}
\seealso{
\itemize{
\item \code{\link{InapplicableFitch}}, calculates parsimony score, supports inapplicable tokens;
\item \code{\link{RootedNNI}}, conducts tree rearrangements;
\item \code{\link{SectorialSearch}}, alternative heuristic, useful for larger trees;
\item \code{\link{Ratchet}}, alternative heuristic, useful to escape local optima.
}
}
\author{
Martin R. Smith
}
\keyword{tree}
