% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{VisualiseInheritance}
\alias{VisualiseInheritance}
\title{Visualize Inheritance}
\usage{
VisualiseInheritance(tree, data, char.no, plot.fun = plot)
}
\arguments{
\item{PARAM}{is a parameter you should send to it}
}
\value{
This function returns :
}
\description{
\code{VisualiseInheritance} does something useful
}
\examples{
to_do <- TRUE

}
\author{
Martin R. Smith
}
