% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{VisualizeCharacter}
\alias{VisualizeCharacter}
\alias{VisualiseCharacter}
\title{VisualiseCharacter}
\usage{
VisualiseCharacter(tree, dataset, char.no, plot.fun = plot, inherit.ancestral = FALSE)
}
\arguments{
\item{tree}{a fully-resolved tree in \code{\link{phylo}} format, with the desired outgroup; edge lengths are not supported and will be deleted;}

\item{dataset}{A phylogenetic data matrix of class \code{\link[phangorn]{phyDat}}, whose names 
correspond to the labels of any accompanying tree}

\item{char.no}{number of the character to be displayed;}

\item{plot.fun}{a function that plots a tree, \code{\link{plot}} by default.}
}
\value{
The function plots a cladogram, annotating each node with the values that could occur there at the lowest parsimony score.
  Nodes highlighed in red potentially (but do not necessarily) contribute to tree length.
}
\description{
Determine and depict the possible states for a character on a tree under the most parsimonious conditions
}
\details{
Visualize optimization for a character
}
\examples{
{
data('SigSut')
outgroup <- c('Lingula', 'Mickwitzia', 'Neocrania')
njtree <- root(nj(dist.hamming(SigSut.phy)), outgroup, resolve.root=TRUE)
njtree$edge.length <- NULL; njtree<-SetOutgroup(njtree, outgroup)
VisualiseCharacter(njtree, SigSut.phy, 5, function(x) {plot(x); edgelabels();})
}
}
\author{
Martin R. Smith
}
\keyword{tree}
