% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_search.R
\name{BootstrapTree}
\alias{BootstrapTree}
\title{Bootstrap tree search with inapplicable data}
\usage{
BootstrapTree(tree, morphyObj, maxiter, maxhits, criterion = criterion,
  verbosity = 1, ...)
}
\arguments{
\item{tree}{A tree of class \code{\link{phylo}}, with tip labels in the order generated by 
\code{\link{RenumberTips}}, i.e. corresponding corresponding the sequence of taxa 
in the corresponding Morphy object.}

\item{morphyObj}{A morphy object created with \code{\link{LoadMorphy}}}
}
\value{
A tree that is optimal under a random sampling of the original characters
}
\description{
Bootstrap tree search with inapplicable data
}
