% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_search.R
\name{DoTreeSearch}
\alias{DoTreeSearch}
\title{DoTreeSearch}
\usage{
DoTreeSearch(tree, morphyObj, Rearrange = TreeSearch::RootedNNI,
  maxIter = 100, maxHits = 20, forestSize = 1, cluster = NULL,
  verbosity = 1, ...)
}
\arguments{
\item{tree}{A tree of class \code{\link{phylo}}, with tip labels in the order generated by 
\code{\link{RenumberTips}}, i.e. corresponding corresponding the sequence of taxa 
in the corresponding Morphy object.}

\item{morphyObj}{A morphy object created with \code{\link{LoadMorphy}}}

\item{Rearrange}{Function to use to rearrange trees; default: \code{\link{RootedNNI}}}

\item{maxIter}{maximum iterations to conduct}

\item{maxHits}{stop search after this many hits}

\item{forestSize}{how many trees to hold}

\item{cluster}{Not really implemented; #TODO}

\item{verbosity}{Level of detail to display in console: larger numbers provides more verbose feedback to the user;}

\item{\dots}{additional variables to pass to RearrangeTree}
}
\description{
Performs a tree search
}
\details{
Does the hard work of searching for a most parsimonious tree.
End-users are expected to access this function through its wrapper, TreeSearch
It is also called directly by Ratchet and Sectorial functions
}
\author{
Martin R. Smith
}
\keyword{internal}
