% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_search.R
\name{Ratchet}
\alias{Ratchet}
\title{Parsimony Ratchet}
\usage{
Ratchet(tree, dataset, concavity = NULL, keepAll = FALSE, outgroup = NULL, maxit = 100, 
               maxiter = 5000, maxhits = 40, k = 10, verbosity = 0, 
               rearrangements = c('TBR', 'SPR', 'NNI'), concavity = NULL, ...)
}
\arguments{
\item{tree}{A tree of class \code{\link{phylo}}}

\item{dataset}{A phylogenetic data matrix of class \code{\link[phangorn]{phyDat}}, whose names 
correspond to the labels of any accompanying tree}

\item{outgroup}{a vector specifying all tips in the outgroup; if unspecified then identical trees with different roots will be considered unique;}

\item{maxit}{maximum ratchet iterations to perform;}

\item{maxiter}{maximum rearrangements to perform on each bootstrap or ratchet iteration;}

\item{maxhits}{maximum times to hit best score before terminating a tree search within a pratchet iteration;}

\item{k}{stop when k ratchet iterations have found the same best score;}

\item{verbosity}{larger numbers provides more verbose feedback to the user;}

\item{rearrangements}{method(s) to use when rearranging trees: 
a vector containing any combination of the strings "NNI", "SPR" or "TBR";}

\item{\dots}{other arguments to pass to subsequent functions.}

\item{concavity}{concavity constant for implied weighting (not currently implemented!)}

\item{all}{Set to \code{TRUE} to report all MPTs encountered during the search, perhaps to analyze consensus}
}
\value{
This function returns a tree modified by parsimony ratchet iteration, retaining the position of the root.
}
\description{
\code{Ratchet} uses the parsimony ratchet (Nixon 1999) to search for a more parsimonious tree.
}
\examples{
{
library('ape')
data('SigSut')
outgroup <- c('Lingula', 'Mickwitzia', 'Neocrania')
njtree <- Root(nj(dist.hamming(SigSut.phy)), outgroup)
njtree$edge.length <- NULL; njtree<-SetOutgroup(njtree, outgroup)
Ratchet(njtree, SigSut.phy, outgroup, maxit=1, maxiter=50)
}
}
\references{
Nixon, K. C. (1999). \cite{The Parsimony Ratchet, a new method for rapid parsimony analysis.} Cladistics, 15(4), 407-414. doi:\href{http://dx.doi.org/10.1111/j.1096-0031.1999.tb00277.x}{10.1111/j.1096-0031.1999.tb00277.x}
}
\seealso{
\code{\link{pratchet}}

\code{\link{TreeSearch}}

\code{\link{SectorialSearch}}
}
\author{
Martin R. Smith

Adapted from \code{\link[phangorn]{pratchet}} in the \pkg{phangorn} package, which does not preserve the position of the root.
}
\keyword{tree}
